Copy all content to middleman site
The header, footer, and sidebar were made partials instead of using
React to create them. The search was removed for now due to trouble
setting it up, but without actual API docs it isn't necessary yet. We
can add that back in once we are generating real docs from source.