Configure pipeline
JavaScript and CSS assets are built with rollup. The `sassc`
gem (libsass-ruby) caused server errors and it was pretty trivial to
setup rollup-plugin-sass.

Formatting helpers were moved to a middleman extension to keep the
config file clean. Another middleman extension was created to add
bigtest package docs to middleman's data.